refactor(core-api): move routes and auth middleware to centralized folders
