+++ apps/core-api/src/routes/auth.routes.ts
