+++ apps/core-api/src/routes/availability.routes.ts
