+++ apps/core-api/src/routes/bookings.routes.ts
