+++ apps/core-api/src/routes/search.routes.ts
