+++ apps/core-api/src/services/auth.service.ts
